{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/requirements/local.txt
-r common.txt

django-debug-toolbar==3.2.1
fabric==2.6.0
